Add support for PureBox
Drawing boxes with ffmpeg filters on the MPV window has many drawbacks,
it is slow, and resource intensive. With this commit I am adding support
for an alternative of drawing the cropbox: PureBox, written with Python
and Xlib. PureBox can be installed with pip install purebox, and
activated in the config file writing the following (it is disabled by
default):

pure_box=yes