Add preliminary support for stream media links
Added support for stream links (YouTube). Video and audio stream links
are extracted and inserted in the copied string. Input seeking works
perfectly, it is very fast which is optimal for long streams.

Switched printf to echo -n to avoid printf's formatting. When Pure Mode
is off, paths will be copied as normal, with no urls extracted.