Add copy_mode
In addition to being able to copy a string for running ffmpeg, now it is
possible to copy a string for purewebm, setting copy_mode to purewebm in
the configuration file. The string will be copied without extra
parameters.

For example, a string that would be copied for ffmpeg with '-c:a copy'
as extra ffmpeg_params like:

ffmpeg -i "/tmp/Videos/test.mkv" -lavfi crop=447:431:195:154 -c:a copy

Will be copied with copy_mode set to purewebm as:

purewebm -i "/tmp/Videos/test.mkv" -lavfi crop=447:431:195:154

Even if purewebm_params is set.