fix: check the status of the xclip execution
Check that xclip was successfully executed before before printing that
the string was copied to the screen. Prints an error if xclip was not
executed.

This fixes an issue where if xclip was not yet installed, the script
would report that the string was copied when in reality it was not.