fix: use ass-events to draw the cropbox
Instead of using ffmpeg filters to draw the cropbox, use ass-events with
mp.create_osd_overlay. This greatly improves performance and works
regardless of which hwdec is set to.

Previously, having the cropbox animation enabled could cause some
undesired slowness, because of this it was disabled by default. It is
now enabled, since this is no longer the case after this commit.