feat: add shared data
With this functionality it is possible to access PureMPV's data with
external scripts using the `user-data` property.

The data can be accessed using the path `user-data/PureMPV`, with
`mp.get_property_native` for example, and it will return an object
containing two properties, `cropbox` and `timestamps`.

The cropbox contains the cropping coordinates `w, h, x, y`, all of type
number if set, null otherwise. The timestamps contains `start` and `end`
of type string in the format HH:MM:SS or null if not yet set.